API SilverStripe 4 compatibility. Add upgrade map, update readme, remove parsedown and include with composer
