+++ .gitattributes
@@ -1,0 +1,1 @@
+/.gitignore export-ignore
