+++ .upgrade.yml
@@ -1,0 +1,3 @@
+mappings:
+  MarkdownEditor: Axllent\Gfmarkdown\Forms\MarkdownEditor
+  Markdown: Axllent\Gfmarkdown\Model\FieldTypes\Markdown
